+++ tests/app/timer/timer-tests.ts
@@ -72,6 +72,23 @@ export function test_setTimeout_callbackCalledAfterSpecifiedTime() {
     TKUnit.assert(completed, "Callback should be called after the specified time!");
 }
 
+export function test_setTimeout_callbackCalledWithBooleanPeriod() {
+    let completed = false;
+
+    // >> timer-set-false
+    const id = timer.setTimeout(() => {
+        // >> (hide)
+        completed = true;
+        // << (hide)
+    // @ts-ignore
+    }, false);
+    // << timer-set-false
+
+    TKUnit.waitUntilReady(() => completed, 1);
+    timer.clearTimeout(id);
+    TKUnit.assert(completed, "Callback should be called in 0 seconds!");
+}
+
 export function test_setTimeout_callbackNotCalled() {
     let completed = false;