+++ tns-core-modules/globals/polyfills/dialogs/dialogs.ts
@@ -1,0 +1,6 @@
+import "../../core";
+import { installPolyfills } from "../polyfill-helpers";
+
+global.registerModule("tns-core-modules/ui/dialogs", () => require("../../../ui/dialogs"));
+
+installPolyfills("tns-core-modules/ui/dialogs", ["alert", "confirm", "prompt", "login", "action"]);